refactor: organize old code into separate packages
