+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/old/PrimaryAuton.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.old;
 
 /* Copyright (c) 2017 FIRST. All rights reserved.
  *
